requirements.txt
litellm
google-adk
openai
python-docx
PyPDF2
python-dotenv
openpyxl>=3.1.0
python-pptx>=0.6.21